ANN 3i3d - ANN_ALGORITHM_CLASSIFICATION_NETWORK: first compile

+++ ANN/ANNalgorithmMemoryNetworkTraining.cpp
@@ -26,7 +26,7 @@
  * File Name: ANNalgorithmMemoryNetworkTraining.cpp
  * Author: Richard Bruce Baxter - Copyright (c) 2005-2016 Baxter AI (baxterai.com)
  * Project: Artificial Neural Network (ANN)
- * Project Version: 4a3c 02-May-2016
+ * Project Version: 4a3d 02-May-2016
  * Comments:
  *
  *******************************************************************************/
@@ -41,7 +41,7 @@
 #ifdef ANN_ALGORITHM_MEMORY_NETWORK
 
 
-void trainNeuralNetworkMemorySimple(ANNneuron* firstInputNeuron, ANNneuron* firstOutputNeuron, int numberOfInputNeurons, int numberOfOutputNeurons, ANNexperience* firstExperienceInDataSet, long numberOfExperiences)
+void trainNeuralNetworkMemorySimple(ANNneuron* firstInputNeuron, ANNneuron* firstOutputNeuron, long numberOfInputNeurons, long numberOfOutputNeurons, ANNexperience* firstExperienceInDataSet, long numberOfExperiences)
 {
 	/*
 	network structure
@@ -102,7 +102,7 @@ void trainNeuralNetworkMemorySimple(ANNneuron* firstInputNeuron, ANNneuron* firs
 
 
 
-void trainNeuralNetworkMemory(ANNneuron* firstInputNeuron, ANNneuron* firstOutputNeuron, int numberOfInputNeurons, int numberOfOutputNeurons, int maxFolds, ANNexperience* firstExperienceInDataSet, long numberOfExperiences)
+void trainNeuralNetworkMemory(ANNneuron* firstInputNeuron, ANNneuron* firstOutputNeuron, long numberOfInputNeurons, long numberOfOutputNeurons, int maxFolds, ANNexperience* firstExperienceInDataSet, long numberOfExperiences)
 {
 	/*
 	network structure
